Move raise error matcher and custom matchers to spec support matchers
Spec support matchers is autoloaded and this
provides a standard location to conveniently
place such files.
