+++ spec/support/matchers/custom_matchers.rb
@@ -13,9 +13,7 @@ RSpec::Matchers.define :contain_claims do |*expected|
   failure_message do |actual|
     "expected that records:\n\t #{actual.inspect} \n\nwould be equal to records\n\t #{expected.inspect}"
   end
-
 end
-
 
 RSpec::Matchers.define :be_within_seconds_of do |expected_date, leeway|
   match do |actual|
